Move prop count check mid-hook to earlier in function
On Windows this hook overrode instructions at the start of the block, which is a jump target for the extension. This made it jump to invalid binary since the real instructions were moved to the trampoline. There is a debug print block right before it, which functionally doesn't do anything for us, so we can hook there instead.